requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-multipart==0.0.9
numpy==2.2.6
pillow==11.1.0
pydicom==3.0.1
opencv-python-headless==4.12.0.88
ultralytics==8.3.248
huggingface-hub==0.27.1
